Have proper reverse optical flow determination, may improve results a little.
The best I can do is increase pyramid levels to two, which makes the lower left hand track.  The legs on the right don't track, I think it is because the whites are saturated and therefore don't have good texture in which to find the flow- some really advance occlusion based tracking could note the texture in the backward is being covered and reveal and infer the solid colored interior is moving, but Farneback does not.  I'd really like to get or implement the TV-L1 flow used by slowmovideo for apples-apples comparison to see if calcOpticalFlowFarneback is doing better or worse, and if the source video is inherently challenging.