Still haven't gotten to root of feedback problem, but changed the way dirtiness propagates from connectors to nodes- it's more instantaneous now rather than happening in the node update.
Also eliminated make a source node dirty just because it connected to a new output.

The basic propagation is that the Output will update the in Connector, which first updates the src Connector on the node, which first requests the parent Node update, the parent then attempts to update all the ports (which may reach back to other Nodes, in this example there is nothing to update, none of them are dirty.

Could have special logic recognize the feedback loop and set dirtiness, though it seems like something is wrong with the implementation if it doesn't work on its own.  There is some asymmetry between Nodes setting thing inside connectors (forward propagation) and dst Connectors look to see if there is something dirty to update with (reach back propagation).